+++ src/tests/fixtures/extend_type/fieldAsArrowFunctionLet.invalid.ts
@@ -1,0 +1,9 @@
+/** @gqlType */
+class SomeType {
+  // No fields
+}
+
+/** @gqlField */
+export let greeting = (_: SomeType): string => {
+  return `Hello World`;
+};