+++ src/tests/fixtures/extend_type/fieldAsExportedAsyncArrowFunction.ts
@@ -1,0 +1,9 @@
+/** @gqlType */
+class SomeType {
+  // No fields
+}
+
+/** @gqlField */
+export const greeting = async (_: SomeType): Promise<string> => {
+  return `Hello World`;
+};