+++ src/tests/fixtures/extend_type/fieldAsExportedMultipleVariables.invalid.ts
@@ -1,0 +1,12 @@
+/** @gqlType */
+class SomeType {
+  // No fields
+}
+
+/** @gqlField */
+export const greeting = (_: SomeType): string => {
+    return `Hello World`;
+  },
+  anotherGreeting = (_: SomeType): string => {
+    return `Hello World`;
+  };