+++ src/tests/fixtures/extend_type/fieldAsExportedNothing.invalid.ts
@@ -1,0 +1,7 @@
+/** @gqlType */
+class SomeType {
+  // No fields
+}
+
+/** @gqlField */
+export const greeting;